✨ chore: enhance documentation and docstring updates
Refine AST-based docs generation, update CLI/runtime doc configs.
