+++ packages/core/src/promptrunner.ts
@@ -123,8 +123,8 @@ async function resolveExpansionVars(
  * @param prj The project context providing runtime and configuration.
  * @param template The prompt script template to execute.
  * @param fragment Additional context such as files, workspace files, and metadata.
- * @param options Configuration for generation, including model, trace, output trace, cancellation token, and other generation parameters.
- * @returns A generation result containing execution details, outputs, and potential errors, including status, messages, edits, annotations, and file changes.
+ * @param options Configuration for generation, including model, trace, output trace, cancellation token, stats, and other generation parameters.
+ * @returns A generation result containing execution details, outputs, and potential errors, including status, messages, edits, annotations, file changes, and usage statistics.
  */
 export async function runTemplate(
     prj: Project,
